requirements.txt
asgiref==3.8.1
cachetools==5.5.2
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.4
dj-database-url==3.0.0
Django==5.2.3
django-allauth==65.9.0
django-storages==1.14.6
google-api-core==2.25.1
google-auth==2.40.3
google-cloud-core==2.4.3
google-cloud-secret-manager==2.24.0
google-cloud-storage==3.1.1
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.73.0
grpcio-status==1.73.0
gunicorn==23.0.0
idna==3.10
packaging==25.0
proto-plus==1.26.1
protobuf==6.31.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.1.0
requests==2.32.4
rsa==4.9.1
sqlparse==0.5.3
typing_extensions==4.14.0
urllib3==2.5.0
whitenoise==6.9.0

django-ckeditor
